Refactor reducers package
