+++ packages/reducers/package.json
@@ -13,6 +13,7 @@
   ],
   "dependencies": {
     "@redux-tools/actions": "^0.1.2",
+    "@redux-tools/injectors": "^0.1.0",
     "@redux-tools/namespaces": "^0.1.2",
     "@redux-tools/utils": "^0.1.2",
     "ramda": "^0.25.0",
