+++ packages/reducers/src/enhancer.js
@@ -1,45 +1,37 @@
-import { forEachObjIndexed, assocPath, keys, dissocPath, forEach, map } from 'ramda';
+import { both, keys, concat, reject } from 'ramda';
+import { createEntries, isVersionEjectable, isEntryIncluded } from '@redux-tools/injectors';
 
 import { reducersInjected, reducersEjected } from './actions';
-import { makeRootReducer, getReducerPath, filterReducer } from './reducers';
+import combineReducerEntries from './combineReducerEntries';
 
 export default function enhancer() {
 	return createStore => (...args) => {
 		const store = createStore(...args);
 
-		store.injectedReducers = {
-			namespaces: {},
+		store.reducerEntries = [];
+
+		store.injectReducers = (reducers, namespace, version) => {
+			store.reducerEntries = concat(
+				store.reducerEntries,
+				createEntries(reducers, namespace, version)
+			);
+
+			store.replaceReducer(combineReducerEntries(store.reducerEntries));
+			store.dispatch(reducersInjected({ reducers: keys(reducers), namespace, version }));
 		};
 
-		const injectReducers = (reducers, namespace) => {
-			const injectReducer = (reducer, key) =>
-				(store.injectedReducers = assocPath(
-					getReducerPath(key, namespace),
-					reducer,
-					store.injectedReducers
-				));
+		store.ejectReducers = (reducers, namespace, version) => {
+			store.reducerEntries = reject(
+				both(
+					isVersionEjectable(version),
+					isEntryIncluded(createEntries(reducers, namespace, version))
+				),
+				store.reducerEntries
+			);
 
-			const filteredReducers = map(filterReducer(namespace), reducers);
-			forEachObjIndexed(injectReducer, filteredReducers);
-			store.replaceReducer(makeRootReducer(store.injectedReducers));
-			store.dispatch(reducersInjected({ reducers: keys(filteredReducers), namespace }));
+			store.replaceReducer(combineReducerEntries(store.reducerEntries));
+			store.dispatch(reducersEjected({ reducers: keys(reducers), namespace, version }));
 		};
-
-		const ejectReducers = (reducers, namespace) => {
-			const ejectReducer = key =>
-				(store.injectedReducers = dissocPath(
-					getReducerPath(key, namespace),
-					store.injectedReducers
-				));
-
-			const reducerKeys = keys(reducers);
-			forEach(ejectReducer, reducerKeys);
-			store.replaceReducer(makeRootReducer(store.injectedReducers));
-			store.dispatch(reducersEjected({ reducers: reducerKeys, namespace }));
-		};
-
-		store.injectReducers = injectReducers;
-		store.ejectReducers = ejectReducers;
 
 		return store;
 	};
