+++ packages/reducers/src/index.js
@@ -1,2 +1,3 @@
 export { default as enhancer } from './enhancer';
-export { getStateByAction, getStateByNamespace } from './reducers';
+export { default as getStateByAction } from './getStateByAction';
+export { default as getStateByNamespace } from './getStateByNamespace';
